Desktop/kamal ai trading new/kamal-ai-trading-pro-main/requirements.txt
black==24.1.0
fastapi==0.95.2
joblib==1.3.2
mypy==1.9.0
openai==1.8.5
prometheus_client==0.15.0
pydantic==2.6.0
pytest==7.4.0
python-dotenv==1.0.1
pyyaml==6.0
rich==13.5.2
ruff==0.24.0
scikit-learn==1.3.2
textual==0.29.0
uvicorn==0.23.0
